backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP y APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Pagos
paypalrestsdk==1.13.1
stripe==7.8.0

# Geolocalización
geopy==2.4.1

# Notificaciones
firebase-admin==6.2.0
twilio==8.10.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoreo y logging
structlog==23.2.0

# Binance Pay y QR codes
sendgrid==6.10.0
qrcode[pil]==7.4.2
Pillow==10.1.0
cryptography==42.0.8 